+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 ### Fixed
 
 1. Error handling for Pushover now works.
+1. Pushover messages are checked for length and split up into multiple messages if necessary.
 
 ## [1.0.0] - 2021-05-17